删除无用的旧版controller目录
