--- app/src/main/kotlin/dev/aaa1115910/bv/component/controllers/GoBackHistoryTip.kt
@@ -1,31 +1,0 @@
-package dev.aaa1115910.bv.component.controllers
-
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import androidx.tv.material3.MaterialTheme
-import androidx.tv.material3.Surface
-import androidx.tv.material3.SurfaceDefaults
-import androidx.tv.material3.Text
-import dev.aaa1115910.bv.util.formatHourMinSec
-
-@Composable
-fun GoBackHistoryTip(
-    modifier: Modifier = Modifier,
-    played: Int
-) {
-    Surface(
-        modifier = modifier,
-        colors = SurfaceDefaults.colors(
-            containerColor = Color.Black.copy(alpha = 0.6f)
-        )
-    ) {
-        Text(
-            modifier = Modifier.padding(8.dp),
-            text = "上次播放到 “${played.toLong().formatHourMinSec()}”，按下确认键返回",
-            style = MaterialTheme.typography.titleLarge
-        )
-    }
-}
