+++ app/src/main/kotlin/dev/aaa1115910/bv/screen/VideoPlayerV3Screen.kt
@@ -34,12 +34,12 @@ import dev.aaa1115910.biliapi.entity.danmaku.DanmakuMaskFrame
 import dev.aaa1115910.biliapi.entity.video.VideoPage
 import dev.aaa1115910.bv.activities.video.UpInfoActivity
 import dev.aaa1115910.bv.component.DanmakuPlayerCompose
-import dev.aaa1115910.bv.component.controllers.LocalVideoPlayerControllerData
-import dev.aaa1115910.bv.component.controllers.VideoPlayerControllerData
-import dev.aaa1115910.bv.component.controllers.info.VideoPlayerInfoData
-import dev.aaa1115910.bv.component.controllers2.DanmakuType
-import dev.aaa1115910.bv.component.controllers2.VideoPlayerController
-import dev.aaa1115910.bv.component.controllers2.VideoProgressSeek
+import dev.aaa1115910.bv.entity.LocalVideoPlayerControllerData
+import dev.aaa1115910.bv.entity.VideoPlayerControllerData
+import dev.aaa1115910.bv.entity.VideoPlayerInfoData
+import dev.aaa1115910.bv.component.controllers.DanmakuType
+import dev.aaa1115910.bv.component.controllers.VideoPlayerController
+import dev.aaa1115910.bv.component.controllers.VideoProgressSeek
 import dev.aaa1115910.bv.component.ifElse
 import dev.aaa1115910.bv.entity.VideoAspectRatio
 import dev.aaa1115910.bv.entity.proxy.ProxyArea
